// backend/src/defect/defect.enums.ts
export enum DefectStatus {
  DRAFT = 'DRAFT',
  ASSIGNED = 'ASSIGNED',
  DIAGNOSIS = 'DIAGNOSIS',
  DIAGNOSIS_DONE = 'DIAGNOSIS_DONE',
  ESTIMATE = 'ESTIMATE',
  APPROVAL = 'APPROVAL',
  APPROVED = 'APPROVED',
  ORDER_CREATED = 'ORDER_CREATED',
  CANCELED = 'CANCELED',
}

export enum DiagnosticStatus {
  NEED_REPLY = 'NEED_REPLY',
  REPLIED = 'REPLIED',
}

export enum ApprovalStatus {
  PENDING = 'PENDING',
  ORDER = 'ORDER',
  RECOMMENDATION = 'RECOMMENDATION',
}